Added Rotations
3 new rotations have been added.

Cardinal: Allows the block to rotate in the 4 cardinal directions.

Pillar: Allows blocks like logs to have the top of the block rotate towards the normal of the block you click on.

Full: Combines cardinal direction with ability to place on the top half of a block to flip it upside down. Useful for stairs.